git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@1023 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/fractalgeneratorservice.cc
@@ -61,7 +61,7 @@ void FractalGeneratorServer::fractalGeneratorPrintParameters(const void* userDat
    const FractalGeneratorServerSettings* settings = (const FractalGeneratorServerSettings*)userData;
 
    puts("Fractal Generator Parameters:");
-   printf("   Failure After = %u [Packets]\n", settings->FailureAfter);
+   printf("   Failure After = %u [Packets]\n", (unsigned int)settings->FailureAfter);
    printf("   Test Mode     = %s\n", (settings->TestMode == true) ? "on" : "off");
 }
